refactor: poem utils
