+++ .changeset/sour-rules-reply.md
@@ -1,0 +1,7 @@
+---
+'neck-upgrade': patch
+---
+
+诗词优化
+
+1. 优化重构获取诗词脚本
