+++ package.json
@@ -11,8 +11,9 @@
     "lint:style": "stylelint \"src/**/*.{css,scss}\"",
     "format": "prettier --write \"src/**/*.{ts,tsx,css,scss}\" \"scripts/**/*.{ts,tsx}\"",
     "prepare": "husky install",
-    "fetch-poetry": "tsx scripts/fetch-poetry/index.ts",
-    "cleanup-poetry": "tsx scripts/fetch-poetry/cleanup.ts",
+    "poetry:download": "tsx scripts/fetch-poetry/index.ts -- download",
+    "poetry:process": "tsx scripts/fetch-poetry/index.ts -- process",
+    "poetry:cleanup": "tsx scripts/fetch-poetry/index.ts -- cleanup",
     "test": "vitest",
     "test:watch": "vitest watch",
     "test:coverage": "vitest run --coverage",
