--- scripts/fetch-poetry/cleanup.ts
@@ -1,3 +1,0 @@
-import { cleanup } from './utils';
-
-cleanup();
